requirements.txt
# ═══════════════════════════════════════
# Job Hunter AI — Python Dependencies
# ═══════════════════════════════════════

# ─── Web Framework ───
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.17

# ─── Database ───
sqlalchemy>=2.0.36

# ─── Authentication ───
bcrypt>=4.2.0
python-jose[cryptography]>=3.3.0

# ─── LLM Integration ───
openai>=1.58.1

# ─── HTTP Client (scraper) ───
httpx[http2]>=0.28.0
tenacity>=9.0.0

# ─── Data Validation ───
pydantic>=2.10.0

# ─── Scheduling ───
apscheduler>=3.10.0

# ─── File Processing ───
python-docx>=1.1.0
PyPDF2>=3.0.0

# ─── Environment ───
python-dotenv>=1.0.0

# ─── Development & Testing ───
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
httpx  # also used by FastAPI TestClient
